Relocate implementations of the public API to single files.
When the `rules_flex` project was first started, the recommended structure
for Bazel rulesets was to have most (or all) of the public API exposed
in a single entry-point file (here, `//flex:flex.bzl`).

Nowadays it's considered better to have the public API in multiple
files, such that the `load()` graph for any given public symbol can be
minimized.

The README will continue to reference the unified `//flex:flex.bzl`
since it's a lot easier to read than a bunch of lengthy `load()` calls.

GitOrigin-RevId: d18a2ff183fd9ff845a960f051e6f947abca722f