feat: backport flex noline fix
When using the --noline flag or the %option noline, flex was failing to
suppress the %top line directives.

There was an if-guard missing in the code that adds the line directives
to the top_buf. I've fixed it in [flex PR #704].

This commit adds patching capabilities to ease maintaining fixes /
backports like the noline fix.

[flex PR #704]: https://github.com/westes/flex/pull/704